Finish 32-bit vmcs control fields
This updates the 32-bit vmcs control fields to use
std::enable_if and bitmanip.h.
