+++ bfvmm/include/vmcs/vmcs_intel_x64_16bit_host_state_fields.h
@@ -54,8 +54,7 @@ namespace host_es_selector
     { return get_vmcs_field_if_exists(addr, name, verbose, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
-    void set(T val)
-    { set_vmcs_field(val, addr, name, exists()); }
+    void set(T val) { set_vmcs_field(val, addr, name, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
     void set_if_exists(T val, bool verbose = false) noexcept
@@ -156,8 +155,7 @@ namespace host_cs_selector
     { return get_vmcs_field_if_exists(addr, name, verbose, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
-    void set(T val)
-    { set_vmcs_field(val, addr, name, exists()); }
+    void set(T val) { set_vmcs_field(val, addr, name, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
     void set_if_exists(T val, bool verbose = false) noexcept
@@ -258,8 +256,7 @@ namespace host_ss_selector
     { return get_vmcs_field_if_exists(addr, name, verbose, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
-    void set(T val)
-    { set_vmcs_field(val, addr, name, exists()); }
+    void set(T val) { set_vmcs_field(val, addr, name, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
     void set_if_exists(T val, bool verbose = false) noexcept
@@ -360,8 +357,7 @@ namespace host_ds_selector
     { return get_vmcs_field_if_exists(addr, name, verbose, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
-    void set(T val)
-    { set_vmcs_field(val, addr, name, exists()); }
+    void set(T val) { set_vmcs_field(val, addr, name, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
     void set_if_exists(T val, bool verbose = false) noexcept
@@ -462,8 +458,7 @@ namespace host_fs_selector
     { return get_vmcs_field_if_exists(addr, name, verbose, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
-    void set(T val)
-    { set_vmcs_field(val, addr, name, exists()); }
+    void set(T val) { set_vmcs_field(val, addr, name, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
     void set_if_exists(T val, bool verbose = false) noexcept
@@ -564,8 +559,7 @@ namespace host_gs_selector
     { return get_vmcs_field_if_exists(addr, name, verbose, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
-    void set(T val)
-    { set_vmcs_field(val, addr, name, exists()); }
+    void set(T val) { set_vmcs_field(val, addr, name, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
     void set_if_exists(T val, bool verbose = false) noexcept
@@ -666,8 +660,7 @@ namespace host_tr_selector
     { return get_vmcs_field_if_exists(addr, name, verbose, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
-    void set(T val)
-    { set_vmcs_field(val, addr, name, exists()); }
+    void set(T val) { set_vmcs_field(val, addr, name, exists()); }
 
     template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
     void set_if_exists(T val, bool verbose = false) noexcept
