+++ bfvmm/src/vmcs/src/vmcs_intel_x64.cpp
@@ -25,6 +25,7 @@
 #include <constants.h>
 #include <thread_context.h>
 #include <vmcs/vmcs_intel_x64.h>
+#include <vmcs/vmcs_intel_x64_32bit_control_fields.h>
 #include <vmcs/vmcs_intel_x64_resume.h>
 #include <vmcs/vmcs_intel_x64_promote.h>
 #include <vmcs/vmcs_intel_x64_16bit_host_state_fields.h>
